requirements.txt
# Trading System Dependencies

# API & Data
schwab-py>=1.5.0
yfinance>=0.2.50
ib_insync>=0.9.86
alpaca-py>=0.30.1  # Alpaca Markets API for market data
# thetadata  # ThetaData — using HTTP API directly, pip package yanked from PyPI
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b0
scipy>=1.11.0  # For Greeks calculations (Black-Scholes model)

# Database
duckdb>=1.1.0
pyarrow>=18.0.0

# ML & Feature Engineering
lightgbm>=4.0.0
scikit-learn>=1.3.0
shap>=0.43.0
joblib>=1.3.0
fredapi>=0.5.0

# Async
aiohttp>=3.11.0
nest-asyncio>=1.6.0

# Monitoring
plotly>=5.24.0
dash>=2.18.0
dash-bootstrap-components>=1.5.0
python-telegram-bot>=21.10
rich>=13.7.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=9.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0

# Development
black>=24.10.0
ruff>=0.9.0
mypy>=1.13.0
